Refactor/ routes for moviedetails add: test: Cypress for movie detatils
