+++ cypress/fixtures/movie_details.json
@@ -1,0 +1,16 @@
+{
+  "backdrop_path": "https://image.tmdb.org/t/p/original//2u7zbn8EudG6kLlBzUYqP8RyFU4.jpg",
+  "genre_ids": [
+    "Adventure",
+    "Fantasy",
+    "Action"
+  ],
+  "id": 122,
+  "original_language": "en",
+  "overview": "As armies mass for a final battle that will decide the fate of the world--and powerful, ancient forces of Light and Dark compete to determine the outcome--one member of the Fellowship of the Ring is revealed as the noble heir to the throne of the Kings of Men. Yet, the sole hope for triumph over evil lies with a brave hobbit, Frodo, who, accompanied by his loyal friend Sam and the hideous, wretched Gollum, ventures deep into the very dark heart of Mordor on his seemingly impossible quest to destroy the Ring of Power.​",
+  "popularity": 175.351,
+  "poster_path": "https://image.tmdb.org/t/p/original//rCzpDGLbOoPwLjy3OAm5NUPOTrC.jpg",
+  "release_date": "2003-12-17",
+  "title": "The Lord of the Rings: The Return of the King",
+  "genres": ["Adventure", "Fantasy", "Action"]
+}
