+++ src/Home_page/Home.css
@@ -1,0 +1,28 @@
+* {
+  margin: 0;
+  padding: 0;
+  }
+
+body {
+  background-color: #3e49b1;
+}
+
+header {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 25px;
+}
+
+h1 {
+  color: white;
+  font-size: 56px;
+  letter-spacing: 5px;
+}
+
+p {
+  font-size: 20px;
+  text-align: center;
+  margin-top: 50px;
+  color: white;
+}
